task 1

+++ 0x17-doubly_linked_lists/1-dlistint_len.c
@@ -7,7 +7,7 @@
  */
 size_t dlistint_len(const dlistint_t *h)
 {
-	size_t element = 0;
+	int element = 0;
 
 	while (h)
 	{
